requirements.txt
Flask==2.3.3
pytesseract==0.3.10
Pillow==12.1.1
werkzeug==2.3.7
